requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
numpy>=1.19.5
opencv-python>=4.5.3
Pillow>=8.3.1
ultralytics>=8.0.0
tensorflow>=2.8.0
tensorflow-hub>=0.12.0
python-dotenv>=0.19.0
pydantic>=1.8.2
matplotlib>=3.4.3
scikit-learn>=0.24.2
requests>=2.25.1
tqdm>=4.61.2
torch>=1.9.0
# onnxruntime>=1.8.0
opencv-python>=4.5.3

